Refactor Github API
