+++ test/actions/test_workflow.py
@@ -2,7 +2,7 @@ from gitbugactions.actions.workflow import GitHubWorkflowFactory
 from gitbugactions.actions.java.maven_workflow import MavenWorkflow
 from gitbugactions.actions.python.pytest_workflow import PytestWorkflow
 from gitbugactions.actions.go.go_workflow import GoWorkflow
-from gitbugactions.github_token import GithubToken
+from gitbugactions.github_api import GithubToken
 
 import os
 import pytest
